+++ src/main/java/org/nim/settings/NimSettingsForm.form
@@ -8,7 +8,7 @@
     <properties/>
     <border type="none"/>
     <children>
-      <grid id="a6cf7" layout-manager="GridLayoutManager" row-count="1" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+      <grid id="a6cf7" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
         <margin top="0" left="0" bottom="0" right="0"/>
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
@@ -16,12 +16,6 @@
         <properties/>
         <border type="etched" title="Nim Sdk"/>
         <children>
-          <component id="e69c7" class="com.intellij.ui.components.JBTextField" binding="sdkHome">
-            <constraints>
-              <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="4" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties/>
-          </component>
           <component id="2bc28" class="com.intellij.ui.components.JBLabel" binding="sdkHomeLabel">
             <constraints>
               <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
@@ -30,13 +24,11 @@
               <text value="SDK Home"/>
             </properties>
           </component>
-          <component id="52fb2" class="javax.swing.JButton" binding="sdkHomeLookup">
+          <component id="813ae" class="com.intellij.openapi.ui.TextFieldWithBrowseButton">
             <constraints>
-              <grid row="0" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+              <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
             </constraints>
-            <properties>
-              <text value="..."/>
-            </properties>
+            <properties/>
           </component>
         </children>
       </grid>
