+++ src/main/java/org/nim/stubs/NimStubFactory.java
@@ -1,7 +1,7 @@
 package org.nim.stubs;
 
 import com.intellij.psi.stubs.IStubElementType;
-import org.nim.stubs.impl.NimMethodLikeStub;
+import org.nim.stubs.impl.NimMethodLikeDeclarationStub;
 import org.nim.stubs.impl.NimTypeStructureStub;
 
 public abstract class NimStubFactory {
@@ -9,8 +9,8 @@ public abstract class NimStubFactory {
     public static IStubElementType<?, ?> build(String name){
         if(NimTypeStructureStub.NAME.equals(name)){
             return NimTypeStructureStub.ADAPTER;
-        } else if(NimMethodLikeStub.NAME.equals(name)){
-            return NimMethodLikeStub.ADAPTER;
+        } else if(NimMethodLikeDeclarationStub.NAME.equals(name)){
+            return NimMethodLikeDeclarationStub.ADAPTER;
         }
         return null;
     }
