+++ src/main/resources/META-INF/plugin.xml
@@ -34,7 +34,7 @@
         <lang.syntaxHighlighterFactory language="Nim" implementationClass="org.nim.color.NimSyntaxHighlighterFactory"/>
         <colorSettingsPage implementation="org.nim.color.NimColorSettingsPage"/>
         <annotator language="Nim" implementationClass="org.nim.color.HighlightingAnnotator" />
-        <sdkType implementation="org.nim.sdk.NimSdkType" />
+<!--        <sdkType implementation="org.nim.sdk.NimSdkType" />-->
         <additionalLibraryRootsProvider implementation="org.nim.nimble.library.NimbleAdditionalLibraryRootsProvider" />
         <projectService serviceImplementation="org.nim.psi.NimPsiService" />
         <applicationConfigurable id="Settings.Nim" bundle="messages.ProjectBundle"
